LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.3_Future_Reserved_Words/S7.5.3_A1.10-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected use of reserved word 'export'
S7.5.3_A1.10

PASS Expected parsing failure

TEST COMPLETE
